@@ -5,13 +5,13 @@ author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  permission_handler: ^5.1.0+2
+  permission_handler: ^6.1.1
   vector_math: ^2.0.8
   json_annotation: ^3.0.1
 
