@@ -1,13 +1,28 @@
 import 'package:ar_flutter_plugin/models/ar_anchor.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter/material.dart';
+
+// Type definitions to enforce a consistent use of the API
+typedef AnchorUploadedHandler = void Function(ARAnchor arAnchor);
+typedef AnchorDownloadedHandler = ARAnchor Function(
+    Map<String, dynamic> serializedAnchor);
 
 /// Manages the session configuration, parameters and events of an [ARView]
 class ARAnchorManager {
   /// Platform channel used for communication from and to [ARAnchorManager]
-  MethodChannel _channel;
+  late MethodChannel _channel;
 
   /// Debugging status flag. If true, all platform calls are printed. Defaults to false.
   final bool debug;
+
+  /// Reference to all anchors that are being uploaded to the google cloud anchor API
+  List<ARAnchor> pendingAnchors = [];
+
+  /// Callback that is triggered once an anchor has successfully been uploaded to the google cloud anchor API
+  AnchorUploadedHandler? onAnchorUploaded;
+
+  /// Callback that is triggered once an anchor has successfully been downloaded from the google cloud anchor API and resolved within the current scene
+  AnchorDownloadedHandler? onAnchorDownloaded;
 
   ARAnchorManager(int id, {this.debug = false}) {
     _channel = MethodChannel('aranchors_$id');
@@ -17,7 +32,11 @@ class ARAnchorManager {
     }
   }
 
-  Future<void> _platformCallHandler(MethodCall call) {
+  initGoogleCloudAnchorMode() async {
+    _channel.invokeMethod<bool>('initGoogleCloudAnchorMode', {});
+  }
+
+  Future<dynamic> _platformCallHandler(MethodCall call) async {
     if (debug) {
       print('_platformCallHandler call ${call.method} ${call.arguments}');
     }
@@ -26,19 +45,44 @@ class ARAnchorManager {
         case 'onError':
           print(call.arguments);
           break;
+        case 'onCloudAnchorUploaded':
+          final name = call.arguments["name"];
+          final cloudanchorid = call.arguments["cloudanchorid"];
+          print(
+              "UPLOADED ANCHOR WITH ID: " + cloudanchorid + ", NAME: " + name);
+          final currentAnchor =
+              pendingAnchors.where((element) => element.name == name).first;
+          // Update anchor with cloud anchor ID
+          (currentAnchor as ARPlaneAnchor).cloudanchorid = cloudanchorid;
+          // Remove anchor from list of pending anchors
+          pendingAnchors.remove(currentAnchor);
+          // Notify callback
+          if (onAnchorUploaded != null) {
+            onAnchorUploaded!(currentAnchor);
+          }
+          break;
+        case "onAnchorDownloadSuccess":
+          final serializedAnchor = call.arguments;
+          if (onAnchorDownloaded != null) {
+            ARAnchor anchor = onAnchorDownloaded!(
+                Map<String, dynamic>.from(serializedAnchor));
+            return anchor.name;
+          } else {
+            return serializedAnchor["name"];
+          }
         default:
           if (debug) {
             print('Unimplemented method ${call.method} ');
           }
       }
     } catch (e) {
-      print('Error caught: ' + e);
+      print('Error caught: ' + e.toString());
     }
     return Future.value();
   }
 
   /// Add given anchor to the underlying AR scene
-  Future<bool> addAnchor(ARAnchor anchor) async {
+  Future<bool?> addAnchor(ARAnchor anchor) async {
     try {
       return await _channel.invokeMethod<bool>('addAnchor', anchor.toJson());
     } on PlatformException catch (e) {
@@ -49,5 +93,23 @@ class ARAnchorManager {
   /// Remove given anchor and all its children from the AR Scene
   removeAnchor(ARAnchor anchor) {
     _channel.invokeMethod<String>('removeAnchor', {'name': anchor.name});
+  }
+
+  /// Upload given anchor from the underlying AR scene to the Google Cloud Anchor API
+  Future<bool?> uploadAnchor(ARAnchor anchor) async {
+    try {
+      final response =
+          await _channel.invokeMethod<bool>('uploadAnchor', anchor.toJson());
+      pendingAnchors.add(anchor);
+      return response;
+    } on PlatformException catch (e) {
+      return false;
+    }
+  }
+
+  Future<bool?> downloadAnchor(String cloudanchorid) async {
+    print("TRYING TO DOWNLOAD ANCHOR WITH ID " + cloudanchorid);
+    _channel
+        .invokeMethod<bool>('downloadAnchor', {"cloudanchorid": cloudanchorid});
   }
 }
