@@ -7,9 +7,9 @@ import 'package:flutter/widgets.dart';
 /// Object specifying a position and rotation in the AR environment
 abstract class ARAnchor {
   ARAnchor({
-    this.type,
-    this.transformation,
-    String name,
+    required this.type,
+    required this.transformation,
+    String? name,
   }) : name = name ?? UniqueKey().toString();
 
   /// Specifies the [AnchorType] of this [ARAnchor]
@@ -22,7 +22,7 @@ abstract class ARAnchor {
   factory ARAnchor.fromJson(Map<String, dynamic> arguments) {
     final type = arguments['type'];
     switch (type) {
-      case AnchorType.plane:
+      case 0: //(= AnchorType.plane)
         return ARPlaneAnchor.fromJson(arguments);
     }
     return ARUnkownAnchor.fromJson(arguments);
@@ -36,14 +36,26 @@ abstract class ARAnchor {
 
 /// An [ARAnchor] fixed to a tracked plane
 class ARPlaneAnchor extends ARAnchor {
-  ARPlaneAnchor(
-      {@required Matrix4 transformation, String name, List<String> childNodes})
-      : childNodes = childNodes ?? [],
+  ARPlaneAnchor({
+    required Matrix4 transformation,
+    String? name,
+    List<String>? childNodes,
+    String? cloudanchorid,
+    int? ttl,
+  })  : childNodes = childNodes ?? [],
+        cloudanchorid = cloudanchorid ?? null,
+        ttl = ttl ?? 1,
         super(
             type: AnchorType.plane, transformation: transformation, name: name);
 
   /// Names of ARNodes attached to this [APlaneRAnchor]
   List<String> childNodes;
+
+  /// ID associated with the anchor after uploading it to the google cloud anchor API
+  String? cloudanchorid;
+
+  /// Time to live of the anchor: Determines how long the anchor is stored once it is uploaded to the google cloud anchor API (optional, defaults to 1 day (24hours))
+  int? ttl;
 
   static ARPlaneAnchor fromJson(Map<String, dynamic> json) =>
       aRPlaneAnchorFromJson(json);
@@ -54,9 +66,15 @@ class ARPlaneAnchor extends ARAnchor {
 
 ARPlaneAnchor aRPlaneAnchorFromJson(Map<String, dynamic> json) {
   return ARPlaneAnchor(
-    transformation: const MatrixConverter().fromJson(json['transform'] as List),
+    transformation:
+        const MatrixConverter().fromJson(json['transformation'] as List),
     name: json['name'] as String,
-    childNodes: json['childNodes'] as List<String>,
+    childNodes: json['childNodes']
+        ?.map((child) => child.toString())
+        ?.toList()
+        ?.cast<String>(),
+    cloudanchorid: json['cloudanchorid'] as String?,
+    ttl: json['ttl'] as int?,
   );
 }
 
@@ -66,12 +84,15 @@ Map<String, dynamic> aRPlaneAnchorToJson(ARPlaneAnchor instance) {
     'transformation': MatrixConverter().toJson(instance.transformation),
     'name': instance.name,
     'childNodes': instance.childNodes,
+    'cloudanchorid': instance.cloudanchorid,
+    'ttl': instance.ttl,
   };
 }
 
 /// An [ARAnchor] type that is not supported yet
 class ARUnkownAnchor extends ARAnchor {
-  ARUnkownAnchor({AnchorType type, Matrix4 transformation, String name})
+  ARUnkownAnchor(
+      {required AnchorType type, required Matrix4 transformation, String? name})
       : super(type: type, transformation: transformation, name: name);
 
   static ARUnkownAnchor fromJson(Map<String, dynamic> json) =>
@@ -84,7 +105,8 @@ class ARUnkownAnchor extends ARAnchor {
 ARUnkownAnchor aRUnkownAnchorFromJson(Map<String, dynamic> json) {
   return ARUnkownAnchor(
     type: json['type'],
-    transformation: const MatrixConverter().fromJson(json['transform'] as List),
+    transformation:
+        const MatrixConverter().fromJson(json['transformation'] as List),
     name: json['name'] as String,
   );
 }
