@@ -4,28 +4,20 @@ import android.app.Activity
 import android.app.Application
 import android.content.Context
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.util.Log
-import android.view.GestureDetector
 import android.view.MotionEvent
 import android.view.View
 import android.widget.Toast
-import androidx.annotation.RequiresApi
 import com.google.ar.core.*
 import com.google.ar.core.exceptions.*
-import com.google.ar.sceneform.ArSceneView
-import com.google.ar.sceneform.FrameTime
-import com.google.ar.sceneform.HitTestResult
-import com.google.ar.sceneform.Node
-import com.google.ar.sceneform.AnchorNode
+import com.google.ar.sceneform.*
 import com.google.ar.sceneform.rendering.Material
 import com.google.ar.sceneform.rendering.PlaneRenderer
 import com.google.ar.sceneform.rendering.Texture
-import com.google.ar.sceneform.math.Vector3
-import com.google.ar.sceneform.math.Quaternion
 import io.carius.lars.ar_flutter_plugin.Serialization.deserializeMatrix4
+import io.carius.lars.ar_flutter_plugin.Serialization.serializeAnchor
 import io.carius.lars.ar_flutter_plugin.Serialization.serializeHitResult
 import io.flutter.FlutterInjector
 import io.flutter.embedding.engine.loader.FlutterLoader
@@ -34,10 +26,7 @@ import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.platform.PlatformView
 import java.nio.FloatBuffer
-import java.util.*
 import java.util.concurrent.CompletableFuture
-import kotlin.collections.ArrayList
-import kotlin.collections.HashMap
 
 internal class AndroidARView(
         val activity: Activity,
@@ -63,6 +52,8 @@ internal class AndroidARView(
     private var worldOriginNode = Node()
     // Model builder
     private var modelBuilder = ArModelBuilder()
+    // Cloud anchor handler
+    private lateinit var cloudAnchorHandler: CloudAnchorHandler
 
     // Method channel handlers
     private val onSessionMethodCall =
@@ -142,8 +133,8 @@ internal class AndroidARView(
                     when (call.method) {
                         "addAnchor" -> {
                             val anchorType: Int? = call.argument<Int>("type")
-                            anchorType?.let{
-                                when(it) {
+                            if (anchorType != null){
+                                when(anchorType) {
                                     0 -> { // Plane Anchor
                                         val transform: ArrayList<Double>? = call.argument<ArrayList<Double>>("transformation")
                                         val name: String? = call.argument<String>("name")
@@ -156,14 +147,49 @@ internal class AndroidARView(
                                     }
                                     else -> result.success(false)
                                 }
-
+                            } else {
+                                result.success(false)
                             }
-                            result.success(false)
                         }
                         "removeAnchor" -> {
                             val anchorName: String? = call.argument<String>("name")
                             anchorName?.let{ name ->
                                 removeAnchor(name)
+                            }
+                        }
+                        "initGoogleCloudAnchorMode" -> {
+                            if (arSceneView.session != null) {
+                                val config = Config(arSceneView.session)
+                                config.cloudAnchorMode = Config.CloudAnchorMode.ENABLED
+                                config.updateMode = Config.UpdateMode.LATEST_CAMERA_IMAGE
+                                config.focusMode = Config.FocusMode.AUTO
+                                arSceneView.session?.configure(config)
+
+                                cloudAnchorHandler = CloudAnchorHandler(arSceneView.session!!)
+                            } else {
+                                sessionManagerChannel.invokeMethod("onError", listOf("Error initializing cloud anchor mode: Session is null"))
+                            }
+                        }
+                        "uploadAnchor" ->  {
+                            val anchorName: String? = call.argument<String>("name")
+                            val ttl: Int? = call.argument<Int>("ttl")
+                            anchorName?.let {
+                                val anchorNode = arSceneView.scene.findByName(anchorName) as AnchorNode?
+                                if (ttl != null) {
+                                    cloudAnchorHandler.hostCloudAnchorWithTtl(anchorName, anchorNode!!.anchor, cloudAnchorUploadedListener(), ttl!!)
+                                } else {
+                                    cloudAnchorHandler.hostCloudAnchor(anchorName, anchorNode!!.anchor, cloudAnchorUploadedListener())
+                                }
+                                //Log.d(TAG, "---------------- HOSTING INITIATED ------------------")
+                                result.success(true)
+                            }
+
+                        }
+                        "downloadAnchor" -> {
+                            val anchorId: String? = call.argument<String>("cloudanchorid")
+                            //Log.d(TAG, "---------------- RESOLVING INITIATED ------------------")
+                            anchorId?.let {
+                                cloudAnchorHandler.resolveCloudAnchor(anchorId, cloudAnchorDownloadedListener())
                             }
                         }
                         else -> {}
@@ -434,6 +460,10 @@ internal class AndroidARView(
             // Release resources
             pointCloud?.release()
         }
+        val updatedAnchors = arSceneView.arFrame!!.updatedAnchors
+        // Notify the cloudManager of all the updates.
+        cloudAnchorHandler.onUpdate(updatedAnchors)
+
     }
 
     private fun addNode(dict_node: HashMap<String, Any>, dict_anchor: HashMap<String, Any>? = null): CompletableFuture<Boolean>{
@@ -447,7 +477,7 @@ internal class AndroidARView(
                     val key: String = loader.getLookupKeyForAsset(dict_node["uri"] as String)
 
                     // Add object to scene
-                    modelBuilder.makeNodeFromGltf(viewContext, dict_node["name"] as String, key, dict_node["transform"] as ArrayList<Double>)
+                    modelBuilder.makeNodeFromGltf(viewContext, dict_node["name"] as String, key, dict_node["transformation"] as ArrayList<Double>)
                             .thenAccept{node ->
                                 val anchorName: String? = dict_anchor?.get("name") as? String
                                 val anchorType: Int? = dict_anchor?.get("type") as? Int
@@ -473,7 +503,7 @@ internal class AndroidARView(
                             }
                 }
                 1 -> { // GLB Model from the web
-                    modelBuilder.makeNodeFromGlb(viewContext, dict_node["name"] as String, dict_node["uri"] as String, dict_node["transform"] as ArrayList<Double>)
+                    modelBuilder.makeNodeFromGlb(viewContext, dict_node["name"] as String, dict_node["uri"] as String, dict_node["transformation"] as ArrayList<Double>)
                             .thenAccept{node ->
                                 val anchorName: String? = dict_anchor?.get("name") as? String
                                 val anchorType: Int? = dict_anchor?.get("type") as? Int
@@ -561,6 +591,56 @@ internal class AndroidARView(
                 // Remove anchor node
                 anchorNode.setParent(null)
             }
+    }
+
+    private inner class cloudAnchorUploadedListener: CloudAnchorHandler.CloudAnchorListener {
+        override fun onCloudTaskComplete(anchorName: String?, anchor: Anchor?) {
+            val cloudState = anchor!!.cloudAnchorState
+            if (cloudState.isError) {
+                Log.e(TAG, "Error uploading anchor, state $cloudState")
+                sessionManagerChannel.invokeMethod("onError", listOf("Error uploading anchor, state $cloudState"))
+                return
+            }
+            // Swap old an new anchor of the respective AnchorNode
+            val anchorNode = arSceneView.scene.findByName(anchorName) as AnchorNode?
+            val oldAnchor = anchorNode?.anchor
+            anchorNode?.anchor = anchor
+            oldAnchor?.detach()
+
+            val args = HashMap<String, String?>()
+            args["name"] = anchorName
+            args["cloudanchorid"] = anchor.cloudAnchorId
+            anchorManagerChannel.invokeMethod("onCloudAnchorUploaded", args)
+        }
+    }
+
+    private inner class cloudAnchorDownloadedListener: CloudAnchorHandler.CloudAnchorListener {
+        override fun onCloudTaskComplete(anchorName: String?, anchor: Anchor?) {
+            val cloudState = anchor!!.cloudAnchorState
+            if (cloudState.isError) {
+                Log.e(TAG, "Error downloading anchor, state $cloudState")
+                sessionManagerChannel.invokeMethod("onError", listOf("Error downloading anchor, state $cloudState"))
+                return
+            }
+            //Log.d(TAG, "---------------- RESOLVING SUCCESSFUL ------------------")
+            val newAnchorNode = AnchorNode(anchor)
+            // Register new anchor on the Flutter side of the plugin
+            anchorManagerChannel.invokeMethod("onAnchorDownloadSuccess", serializeAnchor(newAnchorNode, anchor), object: MethodChannel.Result {
+                override fun success(result: Any?) {
+                    newAnchorNode.name = result.toString()
+                    newAnchorNode.setParent(arSceneView.scene)
+                    //Log.d(TAG, "---------------- REGISTERING ANCHOR SUCCESSFUL ------------------")
+                }
+
+                override fun error(errorCode: String, errorMessage: String?, errorDetails: Any?) {
+                    sessionManagerChannel.invokeMethod("onError", listOf("Error while registering downloaded anchor at the AR Flutter plugin: $errorMessage"))
+                }
+
+                override fun notImplemented() {
+                    sessionManagerChannel.invokeMethod("onError", listOf("Error while registering downloaded anchor at the AR Flutter plugin"))
+                }
+            })
+        }
     }
 
 }
