@@ -3,8 +3,9 @@ import UIKit
 import Foundation
 import ARKit
 import Combine
+import ARCoreCloudAnchors
 
-class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureRecognizerDelegate {
+class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureRecognizerDelegate, ARSessionDelegate {
     let sceneView: ARSCNView
     let sessionManagerChannel: FlutterMethodChannel
     let objectManagerChannel: FlutterMethodChannel
@@ -16,6 +17,10 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
     
     var cancellableCollection = Set<AnyCancellable>() //Used to store all cancellables in (needed for working with Futures)
     var anchorCollection = [String: ARAnchor]() //Used to bookkeep all anchors created by Flutter calls
+    
+    private var cloudAnchorHandler: CloudAnchorHandler? = nil
+    private var arcoreSession: GARSession? = nil
+    private var arcoreMode: Bool = false
 
     init(
         frame: CGRect,
@@ -32,10 +37,12 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
         let configuration = ARWorldTrackingConfiguration() // Create default configuration before initializeARView is called
         self.sceneView.delegate = self
         self.sceneView.session.run(configuration)
+        self.sceneView.session.delegate = self
 
         self.sessionManagerChannel.setMethodCallHandler(self.onSessionMethodCalled)
         self.objectManagerChannel.setMethodCallHandler(self.onObjectMethodCalled)
         self.anchorManagerChannel.setMethodCallHandler(self.onAnchorMethodCalled)
+        
     }
 
     func view() -> UIView {
@@ -122,6 +129,43 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
             case "removeAnchor":
                 if let name = arguments!["name"] as? String {
                     deleteAnchor(anchorName: name)
+                }
+                break
+            case "initGoogleCloudAnchorMode":
+                arcoreSession = try! GARSession.session()
+
+                if (arcoreSession != nil){
+                    if let token = JWTGenerator().generateWebToken(){
+                        arcoreSession!.setAuthToken(token)
+                        
+                        cloudAnchorHandler = CloudAnchorHandler(session: arcoreSession!)
+                        arcoreSession!.delegate = cloudAnchorHandler
+                        arcoreSession!.delegateQueue = DispatchQueue.main
+                        
+                        arcoreMode = true
+                    } else {
+                        sessionManagerChannel.invokeMethod("onError", arguments: ["Error generating JWT, have you added cloudAnchorKey.json into the example/ios/Runner directory?"])
+                    }
+                } else {
+                    sessionManagerChannel.invokeMethod("onError", arguments: ["Error initializing Google AR Session"])
+                }
+                    
+                break
+            case "uploadAnchor":
+                if let anchorName = arguments!["name"] as? String, let anchor = anchorCollection[anchorName] {
+                    print("---------------- HOSTING INITIATED ------------------")
+                    if let ttl = arguments!["ttl"] as? Int {
+                        cloudAnchorHandler?.hostCloudAnchorWithTtl(anchorName: anchorName, anchor: anchor, listener: cloudAnchorUploadedListener(parent: self), ttl: ttl)
+                    } else {
+                        cloudAnchorHandler?.hostCloudAnchor(anchorName: anchorName, anchor: anchor, listener: cloudAnchorUploadedListener(parent: self))
+                    }
+                }
+                result(true)
+                break
+            case "downloadAnchor":
+                if let anchorId = arguments!["cloudanchorid"] as? String {
+                    print("---------------- RESOLVING INITIATED ------------------")
+                    cloudAnchorHandler?.resolveCloudAnchor(anchorId: anchorId, listener: cloudAnchorDownloadedListener(parent: self))
                 }
                 break
             default:
@@ -217,6 +261,16 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
     func renderer(_ renderer: SCNSceneRenderer, didRemove node: SCNNode, for anchor: ARAnchor) {
         trackedPlanes.removeValue(forKey: anchor.identifier)
     }
+    
+    func session(_ session: ARSession, didUpdate frame: ARFrame) {
+        if (arcoreMode) {
+            do {
+                try arcoreSession!.update(frame)
+            } catch {
+                print(error)
+            }
+        }
+    }
 
     func addNode(dict_node: Dictionary<String, Any>, dict_anchor: Dictionary<String, Any>? = nil) -> Future<Bool, Never> {
 
@@ -227,7 +281,7 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
                     // Get path to given Flutter asset
                     let key = FlutterDartProject.lookupKey(forAsset: dict_node["uri"] as! String)
                     // Add object to scene
-                    if let node: SCNNode = self.modelBuilder.makeNodeFromGltf(name: dict_node["name"] as! String, modelPath: key, transformation: dict_node["transform"] as? Array<NSNumber>) {
+                    if let node: SCNNode = self.modelBuilder.makeNodeFromGltf(name: dict_node["name"] as! String, modelPath: key, transformation: dict_node["transformation"] as? Array<NSNumber>) {
                         if let anchorName = dict_anchor?["name"] as? String, let anchorType = dict_anchor?["type"] as? Int {
                             switch anchorType{
                                 case 0: //PlaneAnchor
@@ -253,7 +307,7 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
                     break
                 case 1: // GLB Model from the web
                     // Add object to scene
-                    self.modelBuilder.makeNodeFromWebGlb(name: dict_node["name"] as! String, modelURL: dict_node["uri"] as! String, transformation: dict_node["transform"] as? Array<NSNumber>)
+                    self.modelBuilder.makeNodeFromWebGlb(name: dict_node["name"] as! String, modelURL: dict_node["uri"] as! String, transformation: dict_node["transformation"] as? Array<NSNumber>)
                     .sink(receiveCompletion: {
                                     completion in print("Async Model Downloading Task completed: ", completion)
                     }, receiveValue: { val in
@@ -353,7 +407,92 @@ class IosARView: NSObject, FlutterPlatformView, ARSCNViewDelegate, UIGestureReco
             // Update bookkeeping
             anchorCollection.removeValue(forKey: anchorName)
         }
+    }
+    
+    private class cloudAnchorUploadedListener: CloudAnchorListener {
+        private var parent: IosARView
         
+        init(parent: IosARView) {
+            self.parent = parent
+        }
+        
+        func onCloudTaskComplete(anchorName: String?, anchor: GARAnchor?) {
+            if let cloudState = anchor?.cloudState {
+                if (cloudState == GARCloudAnchorState.success) {
+                    var args = Dictionary<String, String?>()
+                    args["name"] = anchorName
+                    args["cloudanchorid"] = anchor?.cloudIdentifier
+                    parent.anchorManagerChannel.invokeMethod("onCloudAnchorUploaded", arguments: args)
+                } else {
+                    print("Error uploading anchor, state: \(parent.decodeCloudAnchorState(state: cloudState))")
+                    parent.sessionManagerChannel.invokeMethod("onError", arguments: ["Error uploading anchor, state: \(parent.decodeCloudAnchorState(state: cloudState))"])
+                    return
+                }
+            }
+        }
+    }
+
+    private class cloudAnchorDownloadedListener: CloudAnchorListener {
+        private var parent: IosARView
+        
+        init(parent: IosARView) {
+            self.parent = parent
+        }
+        
+        func onCloudTaskComplete(anchorName: String?, anchor: GARAnchor?) {
+            if let cloudState = anchor?.cloudState {
+                if (cloudState == GARCloudAnchorState.success) {
+                    let newAnchor = ARAnchor(transform: anchor!.transform)
+                    // Register new anchor on the Flutter side of the plugin
+                    parent.anchorManagerChannel.invokeMethod("onAnchorDownloadSuccess", arguments: serializeAnchor(anchor: newAnchor, anchorNode: nil, ganchor: anchor!, name: anchorName), result: { result in
+                        if let anchorName = result as? String {
+                            self.parent.sceneView.session.add(anchor: newAnchor)
+                            self.parent.anchorCollection[anchorName] = newAnchor
+                        } else {
+                            self.parent.sessionManagerChannel.invokeMethod("onError", arguments: ["Error while registering downloaded anchor at the AR Flutter plugin"])
+                        }
+
+                    })
+                } else {
+                    print("Error downloading anchor, state \(cloudState)")
+                    parent.sessionManagerChannel.invokeMethod("onError", arguments: ["Error downloading anchor, state \(cloudState)"])
+                    return
+                }
+            }
+        }
+    }
+    
+    func decodeCloudAnchorState(state: GARCloudAnchorState) -> String {
+        switch state {
+        case .errorCloudIdNotFound:
+            return "Cloud anchor id not found"
+        case .errorHostingDatasetProcessingFailed:
+            return "Dataset processing failed, feature map insufficient"
+        case .errorHostingServiceUnavailable:
+            return "Hosting service unavailable"
+        case .errorInternal:
+            return "Internal error"
+        case .errorNotAuthorized:
+            return "Authentication failed: Not Authorized"
+        case .errorResolvingSdkVersionTooNew:
+            return "Resolving Sdk version too new"
+        case .errorResolvingSdkVersionTooOld:
+            return "Resolving Sdk version too old"
+        case .errorResourceExhausted:
+            return " Resource exhausted"
+        case .none:
+            return "Empty state"
+        case .taskInProgress:
+            return "Task in progress"
+        case .success:
+            return "Success"
+        case .errorServiceUnavailable:
+            return "Cloud Anchor Service unavailable"
+        case .errorResolvingLocalizationNoMatch:
+            return "No match"
+        @unknown default:
+            return "Unknown"
+        }
         
     }
         
