@@ -1,5 +1,6 @@
 // The code in this file is adapted from Oleksandr Leuschenko' ARKit Flutter Plugin (https://github.com/olexale/arkit_flutter_plugin)
 
+import 'package:ar_flutter_plugin/utils/json_converters.dart';
 import 'package:flutter/widgets.dart';
 import 'package:vector_math/vector_math_64.dart';
 import 'package:json_annotation/json_annotation.dart';
@@ -11,17 +12,19 @@ import 'package:ar_flutter_plugin/datatypes/node_types.dart';
 /// The coordinate systems of all the sub-nodes are relative to the one of their parent node.
 class ARNode {
   ARNode({
-    this.type,
-    this.uri,
-    String name,
-    Vector3 position,
-    Vector3 scale,
-    Vector4 rotation,
-    Vector3 eulerAngles,
-    Matrix4 transformation,
+    required this.type,
+    required this.uri,
+    String? name,
+    Vector3? position,
+    Vector3? scale,
+    Vector4? rotation,
+    Vector3? eulerAngles,
+    Matrix4? transformation,
+    Map<String, dynamic>? data,
   })  : name = name ?? UniqueKey().toString(),
         transformNotifier = ValueNotifier(createTransformMatrix(
-            transformation, position, scale, rotation, eulerAngles));
+            transformation, position, scale, rotation, eulerAngles)),
+        data = data ?? null;
 
   /// Specifies the receiver's [NodeType]
   NodeType type;
@@ -84,20 +87,34 @@ class ARNode {
   /// Will be autogenerated if not defined.
   final String name;
 
+  /// Holds any data attached to the node, especially useful when uploading serialized nodes to the cloud. This data is not shared with the underlying platform
+  Map<String, dynamic>? data;
+
   static const _matrixValueNotifierConverter = MatrixValueNotifierConverter();
 
   Map<String, dynamic> toMap() => <String, dynamic>{
         'type': type.index,
         'uri': uri,
-        'transform': _matrixValueNotifierConverter.toJson(transformNotifier),
+        'transformation':
+            _matrixValueNotifierConverter.toJson(transformNotifier),
         'name': name,
+        'data': data,
       }..removeWhere((String k, dynamic v) => v == null);
+
+  static ARNode fromMap(Map<String, dynamic> map) {
+    return ARNode(
+        type: NodeType.values[map["type"]],
+        uri: map["uri"] as String,
+        name: map["name"] as String,
+        transformation: MatrixConverter().fromJson(map["transformation"]),
+        data: Map<String, dynamic>.from(map["data"]));
+  }
 }
 
 // Helper functions
 
-Matrix4 createTransformMatrix(Matrix4 origin, Vector3 position, Vector3 scale,
-    Vector4 rotation, Vector3 eulerAngles) {
+Matrix4 createTransformMatrix(Matrix4? origin, Vector3? position,
+    Vector3? scale, Vector4? rotation, Vector3? eulerAngles) {
   final transform = origin ?? Matrix4.identity();
 
   if (position != null) {
@@ -193,11 +210,11 @@ class MatrixValueNotifierConverter
   }
 
   @override
-  List<dynamic> toJson(ValueNotifier<Matrix4> matrix) {
+  List<dynamic>? toJson(ValueNotifier<Matrix4>? matrix) {
     if (matrix == null || matrix.value == null) {
       return null;
     }
-    final list = List<double>(16);
+    final list = List<double>.filled(16, 0.0);
     matrix.value.copyIntoArray(list);
     return list;
   }
